Make symbol types ordered
We want to have symbol types be actual C-like enumerations.
We do not want to mixin 'int' because enum will then act as int for comparisons and in str representation.
For example, when dumping to JSON the parser will treat it as int and print the int value.

We want the enum to act as an enum but also ordering.